too many changes
added new crud operations for projects, so its easier to debug if you can read projects
changed package-structure to ddd and hexagonal architecture
added DTOs for data transfer between application layer and domain layer
added DTO Mapper to convert entities to DTOs more easily
added domain service for uses cases which reffer to business logic (not typical crud. e.g. adding a member to a project or changing the name of the project.
deleted the member repository and the coressponding interfaces, because their is no need. The aggregate root is the project which reffers to each member. Before project and member were two seperate aggregates, which is not correct for our use cases.